api/config/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
pydantic[email]==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
requests==2.31.0 